pyproject.toml
[tool.poetry]
name = "crop-yield-ai"
version = "0.1.0"
description = "Multi-agent AI system for crop yield prediction"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
crewai = "^0.175.0"
langchain = "^0.3.27"
fastapi = "^0.116.1"
uvicorn = "^0.18.3"
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"
spacy = "^3.8.7"
scikit-learn = "^1.7.1"
cryptography = "^45.0.7"
pydantic = "^2.11.7"
streamlit = "^1.49.1"
pandas = "^2.3.2"
python-dotenv = "^1.1.1"
pytest = "^8.4.1"
langchain-community = "^0.3.29"
sentence-transformers = "^5.1.0"
langchain-huggingface = "^0.3.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
